+++ app/Http/Controllers/ProfStandartController.php
@@ -6,43 +6,6 @@ use Illuminate\Http\Request;
 use App\ProfStandart;
 class ProfStandartController extends Controller
 {
-    public function get_profstandarts()
-    {
-        $ps = ProfStandart::orderBy('nameCode')->get();
-        return $ps;
-    }
-
-
-    public function get_profstandart(Request $request)
-    {
-        $ps = ProfStandart::find($request->id);
-        return $ps;
-    }
-
-    public function add_profstandart(Request $request)
-    {
-        $ps = new ProfStandart;
-        $ps->code = $request->code;
-        $ps->name = $request->name;
-        $ps->save();
-        return $ps;
-    }
-
-    public function update_profstandart(Request $request)
-    {
-        $ps = ProfStandart::find($request->id);
-        $ps->code = $request->code;
-        $ps->name = $request->name;
-        $ps->save();
-        return $ps;
-    }
-
-
-    public function remove_profstandart(Request $request)
-    {
-        ProfStandart::destroy($request->id);
-        return $request->id;
-    }
 
     public function index()
     {
@@ -90,7 +53,50 @@ class ProfStandartController extends Controller
 
     public function destroy($id)
     {
+        $ps = ProfStandart::findOrFail($id);
+        if (count($ps->dpps) > 0)
+        {
+            return response()->json(['message'=>'forbidden'],403);
+        }
         ProfStandart::destroy($id);
         return $id;
+    }
+
+    public function get_profstandarts()
+    {
+        $ps = ProfStandart::orderBy('nameCode')->get();
+        return $ps;
+    }
+
+
+    public function get_profstandart(Request $request)
+    {
+        $ps = ProfStandart::find($request->id);
+        return $ps;
+    }
+
+    public function add_profstandart(Request $request)
+    {
+        $ps = new ProfStandart;
+        $ps->code = $request->code;
+        $ps->name = $request->name;
+        $ps->save();
+        return $ps;
+    }
+
+    public function update_profstandart(Request $request)
+    {
+        $ps = ProfStandart::find($request->id);
+        $ps->code = $request->code;
+        $ps->name = $request->name;
+        $ps->save();
+        return $ps;
+    }
+
+
+    public function remove_profstandart(Request $request)
+    {
+        ProfStandart::destroy($request->id);
+        return $request->id;
     }
 }
